+++ tests/generator_cases/optional/expected/optional/optional_runtime_gen.go
@@ -2,6 +2,20 @@
 
 package optional
 
+func zigoMust[T any](value T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return value
+}
+
+func zigoMustMatch[T any](value T, matched bool, err error) (T, bool) {
+	if err != nil {
+		panic(err)
+	}
+	return value, matched
+}
+
 func boolToUint8(value bool) uint8 {
 	if value {
 		return 1