+++ tests/generator_cases/root_constructor/expected/terminal/terminal_runtime_gen.go
@@ -39,3 +39,17 @@ func zigoPoisonAfterPanic(err error, handles ...zigoHandle) error {
 	}
 	return err
 }
+
+func zigoMust[T any](value T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return value
+}
+
+func zigoMustMatch[T any](value T, matched bool, err error) (T, bool) {
+	if err != nil {
+		panic(err)
+	}
+	return value, matched
+}